refactor: extract draft form handling into separate hook
This change extracts the form-related logic from useDraftUtil into a new useDraftForm hook, improving separation of concerns and making the draft utility more modular. The new hook encapsulates form open/close behavior and form props management.
